Fixed delta values for confidence interval calculation.  Some reorganization.

+++ gridlock.h
@@ -21,7 +21,7 @@ typedef struct
   char fitType[256];//the type of fit (linear,parabola,etc)
   long double fitOpt;//fit option value (eg. delta for Deming regression) 
   char dataType[256];//the type of data provided (regular, chisq values, etc.)
-  int ciDelta;//delta value for confidence interval calculation
+  long double ciDelta;//delta value for confidence interval calculation
   int plotData;//0=don't plot, 1=plot
   char plotMode[256];//the plotting style to be used
   int plotCI;//0=don't plot confidence interval, 1=plot it
